requirements.txt
openai>=1.40.0
pillow>=10.3.0
pyyaml>=6.0.1
requests>=2.31.0
